Prevent multiple IN predicates to the same attribute
It will now merge the homogenous IN predicates by intersecting
the values after adding it to the where clause. It will decrease the
size of SQL statement.

Previously, it would cause huge sql statements based on how many times
it performed the where method and the size of the lists.