Prevent multiple IN predicates to the same attribute
It will now merge the homogenous IN predicates by intersecting
the values before adding it to the where clause.

Previously, it would add each IN predicate to the where clause
even if they had the same values. It might cause huge sql
statements based on how many times it gets added and the size of
the lists.